IMCMS-176 - Image does not show if published version: - Fixed changing image process to not loose previous when versions allowed.

+++ src/main/java/com/imcode/imcms/servlet/apis/ImageController.java
@@ -175,7 +175,11 @@ public class ImageController {
                 : null;
 
         if (StringUtils.isNotBlank(imageDomainObject.getGeneratedFilename())) {
-            imageDomainObject.getGeneratedFile().delete();
+            if (Imcms.isVersioningAllowed()) {
+                // todo : rewrite process when versions are turned on to delete old image only if it not used in other versions
+            } else {
+                imageDomainObject.getGeneratedFile().delete();
+            }
         }
 
         imageDomainObject.generateFilename();
